services/identity-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "identity-service"
version = "0.1.0"
description = "Household identity, roles, and permissions (ADR-012)"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.1",
    "structlog>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
